Receive flow UI
