+++ lib/core/di/bloc_module.dart
@@ -107,6 +107,14 @@ void _registerBlocsModule() {
     ),
   );
 
+  _registerFactoryWithParams<ReceiveBloc, WalletTokenData, void>(
+        (tokenData, _) => ReceiveBloc(
+      tokenData,
+      _getIt<SendTokenUseCase>(),
+      _getIt<ErrorHandlerManager>(),
+    ),
+  );
+
   _registerFactory(
     () => PushNotificationsBloc(
       _getIt<FirebasePushNotificationsService>(),
