+++ lib/ui/transfer_tokens/receive/interactor/receive_bloc.freezed.dart
@@ -3,7 +3,7 @@
 // ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
-part of 'send_bloc.dart';
+part of 'receive_bloc.dart';
 
 // **************************************************************************
 // FreezedGenerator
@@ -225,17 +225,14 @@ abstract class _NavigateToSendSuccess implements PageCommand {
 }
 
 /// @nodoc
-mixin _$SendEvent {
+mixin _$ReceiveEvent {
   @optionalTypeArgs
   TResult when<TResult extends Object?>({
     required TResult Function() initial,
     required TResult Function() onRefresh,
     required TResult Function() clearPageCommand,
-    required TResult Function(AmountPercentage amountPercentage)
-        onPercentageTapped,
     required TResult Function(KeypadKey tappedKey) onKeypadTapped,
     required TResult Function(String? memo) onMemoEntered,
-    required TResult Function() onSendTapped,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
@@ -243,10 +240,8 @@ mixin _$SendEvent {
     TResult? Function()? initial,
     TResult? Function()? onRefresh,
     TResult? Function()? clearPageCommand,
-    TResult? Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult? Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult? Function(String? memo)? onMemoEntered,
-    TResult? Function()? onSendTapped,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
@@ -254,10 +249,8 @@ mixin _$SendEvent {
     TResult Function()? initial,
     TResult Function()? onRefresh,
     TResult Function()? clearPageCommand,
-    TResult Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult Function(String? memo)? onMemoEntered,
-    TResult Function()? onSendTapped,
     required TResult orElse(),
   }) =>
       throw _privateConstructorUsedError;
@@ -266,10 +259,8 @@ mixin _$SendEvent {
     required TResult Function(_Initial value) initial,
     required TResult Function(_OnRefresh value) onRefresh,
     required TResult Function(_ClearPageCommand value) clearPageCommand,
-    required TResult Function(_OnPercentageTapped value) onPercentageTapped,
     required TResult Function(_OnKeypadTapped value) onKeypadTapped,
     required TResult Function(_OnMemoEntered value) onMemoEntered,
-    required TResult Function(_OnSendTapped value) onSendTapped,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
@@ -277,10 +268,8 @@ mixin _$SendEvent {
     TResult? Function(_Initial value)? initial,
     TResult? Function(_OnRefresh value)? onRefresh,
     TResult? Function(_ClearPageCommand value)? clearPageCommand,
-    TResult? Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult? Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult? Function(_OnMemoEntered value)? onMemoEntered,
-    TResult? Function(_OnSendTapped value)? onSendTapped,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
@@ -288,25 +277,24 @@ mixin _$SendEvent {
     TResult Function(_Initial value)? initial,
     TResult Function(_OnRefresh value)? onRefresh,
     TResult Function(_ClearPageCommand value)? clearPageCommand,
-    TResult Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult Function(_OnMemoEntered value)? onMemoEntered,
-    TResult Function(_OnSendTapped value)? onSendTapped,
     required TResult orElse(),
   }) =>
       throw _privateConstructorUsedError;
 }
 
 /// @nodoc
-abstract class $SendEventCopyWith<$Res> {
-  factory $SendEventCopyWith(SendEvent value, $Res Function(SendEvent) then) =
-      _$SendEventCopyWithImpl<$Res, SendEvent>;
+abstract class $ReceiveEventCopyWith<$Res> {
+  factory $ReceiveEventCopyWith(
+          ReceiveEvent value, $Res Function(ReceiveEvent) then) =
+      _$ReceiveEventCopyWithImpl<$Res, ReceiveEvent>;
 }
 
 /// @nodoc
-class _$SendEventCopyWithImpl<$Res, $Val extends SendEvent>
-    implements $SendEventCopyWith<$Res> {
-  _$SendEventCopyWithImpl(this._value, this._then);
+class _$ReceiveEventCopyWithImpl<$Res, $Val extends ReceiveEvent>
+    implements $ReceiveEventCopyWith<$Res> {
+  _$ReceiveEventCopyWithImpl(this._value, this._then);
 
   // ignore: unused_field
   final $Val _value;
@@ -323,7 +311,7 @@ abstract class _$$_InitialCopyWith<$Res> {
 
 /// @nodoc
 class __$$_InitialCopyWithImpl<$Res>
-    extends _$SendEventCopyWithImpl<$Res, _$_Initial>
+    extends _$ReceiveEventCopyWithImpl<$Res, _$_Initial>
     implements _$$_InitialCopyWith<$Res> {
   __$$_InitialCopyWithImpl(_$_Initial _value, $Res Function(_$_Initial) _then)
       : super(_value, _then);
@@ -336,7 +324,7 @@ class _$_Initial implements _Initial {
 
   @override
   String toString() {
-    return 'SendEvent.initial()';
+    return 'ReceiveEvent.initial()';
   }
 
   @override
@@ -354,11 +342,8 @@ class _$_Initial implements _Initial {
     required TResult Function() initial,
     required TResult Function() onRefresh,
     required TResult Function() clearPageCommand,
-    required TResult Function(AmountPercentage amountPercentage)
-        onPercentageTapped,
     required TResult Function(KeypadKey tappedKey) onKeypadTapped,
     required TResult Function(String? memo) onMemoEntered,
-    required TResult Function() onSendTapped,
   }) {
     return initial();
   }
@@ -369,10 +354,8 @@ class _$_Initial implements _Initial {
     TResult? Function()? initial,
     TResult? Function()? onRefresh,
     TResult? Function()? clearPageCommand,
-    TResult? Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult? Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult? Function(String? memo)? onMemoEntered,
-    TResult? Function()? onSendTapped,
   }) {
     return initial?.call();
   }
@@ -383,10 +366,8 @@ class _$_Initial implements _Initial {
     TResult Function()? initial,
     TResult Function()? onRefresh,
     TResult Function()? clearPageCommand,
-    TResult Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult Function(String? memo)? onMemoEntered,
-    TResult Function()? onSendTapped,
     required TResult orElse(),
   }) {
     if (initial != null) {
@@ -401,10 +382,8 @@ class _$_Initial implements _Initial {
     required TResult Function(_Initial value) initial,
     required TResult Function(_OnRefresh value) onRefresh,
     required TResult Function(_ClearPageCommand value) clearPageCommand,
-    required TResult Function(_OnPercentageTapped value) onPercentageTapped,
     required TResult Function(_OnKeypadTapped value) onKeypadTapped,
     required TResult Function(_OnMemoEntered value) onMemoEntered,
-    required TResult Function(_OnSendTapped value) onSendTapped,
   }) {
     return initial(this);
   }
@@ -415,10 +394,8 @@ class _$_Initial implements _Initial {
     TResult? Function(_Initial value)? initial,
     TResult? Function(_OnRefresh value)? onRefresh,
     TResult? Function(_ClearPageCommand value)? clearPageCommand,
-    TResult? Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult? Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult? Function(_OnMemoEntered value)? onMemoEntered,
-    TResult? Function(_OnSendTapped value)? onSendTapped,
   }) {
     return initial?.call(this);
   }
@@ -429,10 +406,8 @@ class _$_Initial implements _Initial {
     TResult Function(_Initial value)? initial,
     TResult Function(_OnRefresh value)? onRefresh,
     TResult Function(_ClearPageCommand value)? clearPageCommand,
-    TResult Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult Function(_OnMemoEntered value)? onMemoEntered,
-    TResult Function(_OnSendTapped value)? onSendTapped,
     required TResult orElse(),
   }) {
     if (initial != null) {
@@ -442,7 +417,7 @@ class _$_Initial implements _Initial {
   }
 }
 
-abstract class _Initial implements SendEvent {
+abstract class _Initial implements ReceiveEvent {
   const factory _Initial() = _$_Initial;
 }
 
@@ -455,7 +430,7 @@ abstract class _$$_OnRefreshCopyWith<$Res> {
 
 /// @nodoc
 class __$$_OnRefreshCopyWithImpl<$Res>
-    extends _$SendEventCopyWithImpl<$Res, _$_OnRefresh>
+    extends _$ReceiveEventCopyWithImpl<$Res, _$_OnRefresh>
     implements _$$_OnRefreshCopyWith<$Res> {
   __$$_OnRefreshCopyWithImpl(
       _$_OnRefresh _value, $Res Function(_$_OnRefresh) _then)
@@ -469,7 +444,7 @@ class _$_OnRefresh implements _OnRefresh {
 
   @override
   String toString() {
-    return 'SendEvent.onRefresh()';
+    return 'ReceiveEvent.onRefresh()';
   }
 
   @override
@@ -487,11 +462,8 @@ class _$_OnRefresh implements _OnRefresh {
     required TResult Function() initial,
     required TResult Function() onRefresh,
     required TResult Function() clearPageCommand,
-    required TResult Function(AmountPercentage amountPercentage)
-        onPercentageTapped,
     required TResult Function(KeypadKey tappedKey) onKeypadTapped,
     required TResult Function(String? memo) onMemoEntered,
-    required TResult Function() onSendTapped,
   }) {
     return onRefresh();
   }
@@ -502,10 +474,8 @@ class _$_OnRefresh implements _OnRefresh {
     TResult? Function()? initial,
     TResult? Function()? onRefresh,
     TResult? Function()? clearPageCommand,
-    TResult? Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult? Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult? Function(String? memo)? onMemoEntered,
-    TResult? Function()? onSendTapped,
   }) {
     return onRefresh?.call();
   }
@@ -516,10 +486,8 @@ class _$_OnRefresh implements _OnRefresh {
     TResult Function()? initial,
     TResult Function()? onRefresh,
     TResult Function()? clearPageCommand,
-    TResult Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult Function(String? memo)? onMemoEntered,
-    TResult Function()? onSendTapped,
     required TResult orElse(),
   }) {
     if (onRefresh != null) {
@@ -534,10 +502,8 @@ class _$_OnRefresh implements _OnRefresh {
     required TResult Function(_Initial value) initial,
     required TResult Function(_OnRefresh value) onRefresh,
     required TResult Function(_ClearPageCommand value) clearPageCommand,
-    required TResult Function(_OnPercentageTapped value) onPercentageTapped,
     required TResult Function(_OnKeypadTapped value) onKeypadTapped,
     required TResult Function(_OnMemoEntered value) onMemoEntered,
-    required TResult Function(_OnSendTapped value) onSendTapped,
   }) {
     return onRefresh(this);
   }
@@ -548,10 +514,8 @@ class _$_OnRefresh implements _OnRefresh {
     TResult? Function(_Initial value)? initial,
     TResult? Function(_OnRefresh value)? onRefresh,
     TResult? Function(_ClearPageCommand value)? clearPageCommand,
-    TResult? Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult? Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult? Function(_OnMemoEntered value)? onMemoEntered,
-    TResult? Function(_OnSendTapped value)? onSendTapped,
   }) {
     return onRefresh?.call(this);
   }
@@ -562,10 +526,8 @@ class _$_OnRefresh implements _OnRefresh {
     TResult Function(_Initial value)? initial,
     TResult Function(_OnRefresh value)? onRefresh,
     TResult Function(_ClearPageCommand value)? clearPageCommand,
-    TResult Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult Function(_OnMemoEntered value)? onMemoEntered,
-    TResult Function(_OnSendTapped value)? onSendTapped,
     required TResult orElse(),
   }) {
     if (onRefresh != null) {
@@ -575,7 +537,7 @@ class _$_OnRefresh implements _OnRefresh {
   }
 }
 
-abstract class _OnRefresh implements SendEvent {
+abstract class _OnRefresh implements ReceiveEvent {
   const factory _OnRefresh() = _$_OnRefresh;
 }
 
@@ -588,7 +550,7 @@ abstract class _$$_ClearPageCommandCopyWith<$Res> {
 
 /// @nodoc
 class __$$_ClearPageCommandCopyWithImpl<$Res>
-    extends _$SendEventCopyWithImpl<$Res, _$_ClearPageCommand>
+    extends _$ReceiveEventCopyWithImpl<$Res, _$_ClearPageCommand>
     implements _$$_ClearPageCommandCopyWith<$Res> {
   __$$_ClearPageCommandCopyWithImpl(
       _$_ClearPageCommand _value, $Res Function(_$_ClearPageCommand) _then)
@@ -602,7 +564,7 @@ class _$_ClearPageCommand implements _ClearPageCommand {
 
   @override
   String toString() {
-    return 'SendEvent.clearPageCommand()';
+    return 'ReceiveEvent.clearPageCommand()';
   }
 
   @override
@@ -620,11 +582,8 @@ class _$_ClearPageCommand implements _ClearPageCommand {
     required TResult Function() initial,
     required TResult Function() onRefresh,
     required TResult Function() clearPageCommand,
-    required TResult Function(AmountPercentage amountPercentage)
-        onPercentageTapped,
     required TResult Function(KeypadKey tappedKey) onKeypadTapped,
     required TResult Function(String? memo) onMemoEntered,
-    required TResult Function() onSendTapped,
   }) {
     return clearPageCommand();
   }
@@ -635,10 +594,8 @@ class _$_ClearPageCommand implements _ClearPageCommand {
     TResult? Function()? initial,
     TResult? Function()? onRefresh,
     TResult? Function()? clearPageCommand,
-    TResult? Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult? Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult? Function(String? memo)? onMemoEntered,
-    TResult? Function()? onSendTapped,
   }) {
     return clearPageCommand?.call();
   }
@@ -649,10 +606,8 @@ class _$_ClearPageCommand implements _ClearPageCommand {
     TResult Function()? initial,
     TResult Function()? onRefresh,
     TResult Function()? clearPageCommand,
-    TResult Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult Function(String? memo)? onMemoEntered,
-    TResult Function()? onSendTapped,
     required TResult orElse(),
   }) {
     if (clearPageCommand != null) {
@@ -667,10 +622,8 @@ class _$_ClearPageCommand implements _ClearPageCommand {
     required TResult Function(_Initial value) initial,
     required TResult Function(_OnRefresh value) onRefresh,
     required TResult Function(_ClearPageCommand value) clearPageCommand,
-    required TResult Function(_OnPercentageTapped value) onPercentageTapped,
     required TResult Function(_OnKeypadTapped value) onKeypadTapped,
     required TResult Function(_OnMemoEntered value) onMemoEntered,
-    required TResult Function(_OnSendTapped value) onSendTapped,
   }) {
     return clearPageCommand(this);
   }
@@ -681,10 +634,8 @@ class _$_ClearPageCommand implements _ClearPageCommand {
     TResult? Function(_Initial value)? initial,
     TResult? Function(_OnRefresh value)? onRefresh,
     TResult? Function(_ClearPageCommand value)? clearPageCommand,
-    TResult? Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult? Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult? Function(_OnMemoEntered value)? onMemoEntered,
-    TResult? Function(_OnSendTapped value)? onSendTapped,
   }) {
     return clearPageCommand?.call(this);
   }
@@ -695,10 +646,8 @@ class _$_ClearPageCommand implements _ClearPageCommand {
     TResult Function(_Initial value)? initial,
     TResult Function(_OnRefresh value)? onRefresh,
     TResult Function(_ClearPageCommand value)? clearPageCommand,
-    TResult Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult Function(_OnMemoEntered value)? onMemoEntered,
-    TResult Function(_OnSendTapped value)? onSendTapped,
     required TResult orElse(),
   }) {
     if (clearPageCommand != null) {
@@ -708,175 +657,8 @@ class _$_ClearPageCommand implements _ClearPageCommand {
   }
 }
 
-abstract class _ClearPageCommand implements SendEvent {
+abstract class _ClearPageCommand implements ReceiveEvent {
   const factory _ClearPageCommand() = _$_ClearPageCommand;
-}
-
-/// @nodoc
-abstract class _$$_OnPercentageTappedCopyWith<$Res> {
-  factory _$$_OnPercentageTappedCopyWith(_$_OnPercentageTapped value,
-          $Res Function(_$_OnPercentageTapped) then) =
-      __$$_OnPercentageTappedCopyWithImpl<$Res>;
-  @useResult
-  $Res call({AmountPercentage amountPercentage});
-}
-
-/// @nodoc
-class __$$_OnPercentageTappedCopyWithImpl<$Res>
-    extends _$SendEventCopyWithImpl<$Res, _$_OnPercentageTapped>
-    implements _$$_OnPercentageTappedCopyWith<$Res> {
-  __$$_OnPercentageTappedCopyWithImpl(
-      _$_OnPercentageTapped _value, $Res Function(_$_OnPercentageTapped) _then)
-      : super(_value, _then);
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? amountPercentage = null,
-  }) {
-    return _then(_$_OnPercentageTapped(
-      null == amountPercentage
-          ? _value.amountPercentage
-          : amountPercentage // ignore: cast_nullable_to_non_nullable
-              as AmountPercentage,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$_OnPercentageTapped implements _OnPercentageTapped {
-  const _$_OnPercentageTapped(this.amountPercentage);
-
-  @override
-  final AmountPercentage amountPercentage;
-
-  @override
-  String toString() {
-    return 'SendEvent.onPercentageTapped(amountPercentage: $amountPercentage)';
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$_OnPercentageTapped &&
-            (identical(other.amountPercentage, amountPercentage) ||
-                other.amountPercentage == amountPercentage));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, amountPercentage);
-
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$_OnPercentageTappedCopyWith<_$_OnPercentageTapped> get copyWith =>
-      __$$_OnPercentageTappedCopyWithImpl<_$_OnPercentageTapped>(
-          this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() onRefresh,
-    required TResult Function() clearPageCommand,
-    required TResult Function(AmountPercentage amountPercentage)
-        onPercentageTapped,
-    required TResult Function(KeypadKey tappedKey) onKeypadTapped,
-    required TResult Function(String? memo) onMemoEntered,
-    required TResult Function() onSendTapped,
-  }) {
-    return onPercentageTapped(amountPercentage);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? onRefresh,
-    TResult? Function()? clearPageCommand,
-    TResult? Function(AmountPercentage amountPercentage)? onPercentageTapped,
-    TResult? Function(KeypadKey tappedKey)? onKeypadTapped,
-    TResult? Function(String? memo)? onMemoEntered,
-    TResult? Function()? onSendTapped,
-  }) {
-    return onPercentageTapped?.call(amountPercentage);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? onRefresh,
-    TResult Function()? clearPageCommand,
-    TResult Function(AmountPercentage amountPercentage)? onPercentageTapped,
-    TResult Function(KeypadKey tappedKey)? onKeypadTapped,
-    TResult Function(String? memo)? onMemoEntered,
-    TResult Function()? onSendTapped,
-    required TResult orElse(),
-  }) {
-    if (onPercentageTapped != null) {
-      return onPercentageTapped(amountPercentage);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(_Initial value) initial,
-    required TResult Function(_OnRefresh value) onRefresh,
-    required TResult Function(_ClearPageCommand value) clearPageCommand,
-    required TResult Function(_OnPercentageTapped value) onPercentageTapped,
-    required TResult Function(_OnKeypadTapped value) onKeypadTapped,
-    required TResult Function(_OnMemoEntered value) onMemoEntered,
-    required TResult Function(_OnSendTapped value) onSendTapped,
-  }) {
-    return onPercentageTapped(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(_Initial value)? initial,
-    TResult? Function(_OnRefresh value)? onRefresh,
-    TResult? Function(_ClearPageCommand value)? clearPageCommand,
-    TResult? Function(_OnPercentageTapped value)? onPercentageTapped,
-    TResult? Function(_OnKeypadTapped value)? onKeypadTapped,
-    TResult? Function(_OnMemoEntered value)? onMemoEntered,
-    TResult? Function(_OnSendTapped value)? onSendTapped,
-  }) {
-    return onPercentageTapped?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(_Initial value)? initial,
-    TResult Function(_OnRefresh value)? onRefresh,
-    TResult Function(_ClearPageCommand value)? clearPageCommand,
-    TResult Function(_OnPercentageTapped value)? onPercentageTapped,
-    TResult Function(_OnKeypadTapped value)? onKeypadTapped,
-    TResult Function(_OnMemoEntered value)? onMemoEntered,
-    TResult Function(_OnSendTapped value)? onSendTapped,
-    required TResult orElse(),
-  }) {
-    if (onPercentageTapped != null) {
-      return onPercentageTapped(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class _OnPercentageTapped implements SendEvent {
-  const factory _OnPercentageTapped(final AmountPercentage amountPercentage) =
-      _$_OnPercentageTapped;
-
-  AmountPercentage get amountPercentage;
-  @JsonKey(ignore: true)
-  _$$_OnPercentageTappedCopyWith<_$_OnPercentageTapped> get copyWith =>
-      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -890,7 +672,7 @@ abstract class _$$_OnKeypadTappedCopyWith<$Res> {
 
 /// @nodoc
 class __$$_OnKeypadTappedCopyWithImpl<$Res>
-    extends _$SendEventCopyWithImpl<$Res, _$_OnKeypadTapped>
+    extends _$ReceiveEventCopyWithImpl<$Res, _$_OnKeypadTapped>
     implements _$$_OnKeypadTappedCopyWith<$Res> {
   __$$_OnKeypadTappedCopyWithImpl(
       _$_OnKeypadTapped _value, $Res Function(_$_OnKeypadTapped) _then)
@@ -899,10 +681,10 @@ class __$$_OnKeypadTappedCopyWithImpl<$Res>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
-    Object? tappedKey = null,
+    Object? tappedKey = freezed,
   }) {
     return _then(_$_OnKeypadTapped(
-      null == tappedKey
+      freezed == tappedKey
           ? _value.tappedKey
           : tappedKey // ignore: cast_nullable_to_non_nullable
               as KeypadKey,
@@ -920,7 +702,7 @@ class _$_OnKeypadTapped implements _OnKeypadTapped {
 
   @override
   String toString() {
-    return 'SendEvent.onKeypadTapped(tappedKey: $tappedKey)';
+    return 'ReceiveEvent.onKeypadTapped(tappedKey: $tappedKey)';
   }
 
   @override
@@ -928,12 +710,12 @@ class _$_OnKeypadTapped implements _OnKeypadTapped {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$_OnKeypadTapped &&
-            (identical(other.tappedKey, tappedKey) ||
-                other.tappedKey == tappedKey));
+            const DeepCollectionEquality().equals(other.tappedKey, tappedKey));
   }
 
   @override
-  int get hashCode => Object.hash(runtimeType, tappedKey);
+  int get hashCode =>
+      Object.hash(runtimeType, const DeepCollectionEquality().hash(tappedKey));
 
   @JsonKey(ignore: true)
   @override
@@ -947,11 +729,8 @@ class _$_OnKeypadTapped implements _OnKeypadTapped {
     required TResult Function() initial,
     required TResult Function() onRefresh,
     required TResult Function() clearPageCommand,
-    required TResult Function(AmountPercentage amountPercentage)
-        onPercentageTapped,
     required TResult Function(KeypadKey tappedKey) onKeypadTapped,
     required TResult Function(String? memo) onMemoEntered,
-    required TResult Function() onSendTapped,
   }) {
     return onKeypadTapped(tappedKey);
   }
@@ -962,10 +741,8 @@ class _$_OnKeypadTapped implements _OnKeypadTapped {
     TResult? Function()? initial,
     TResult? Function()? onRefresh,
     TResult? Function()? clearPageCommand,
-    TResult? Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult? Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult? Function(String? memo)? onMemoEntered,
-    TResult? Function()? onSendTapped,
   }) {
     return onKeypadTapped?.call(tappedKey);
   }
@@ -976,10 +753,8 @@ class _$_OnKeypadTapped implements _OnKeypadTapped {
     TResult Function()? initial,
     TResult Function()? onRefresh,
     TResult Function()? clearPageCommand,
-    TResult Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult Function(String? memo)? onMemoEntered,
-    TResult Function()? onSendTapped,
     required TResult orElse(),
   }) {
     if (onKeypadTapped != null) {
@@ -994,10 +769,8 @@ class _$_OnKeypadTapped implements _OnKeypadTapped {
     required TResult Function(_Initial value) initial,
     required TResult Function(_OnRefresh value) onRefresh,
     required TResult Function(_ClearPageCommand value) clearPageCommand,
-    required TResult Function(_OnPercentageTapped value) onPercentageTapped,
     required TResult Function(_OnKeypadTapped value) onKeypadTapped,
     required TResult Function(_OnMemoEntered value) onMemoEntered,
-    required TResult Function(_OnSendTapped value) onSendTapped,
   }) {
     return onKeypadTapped(this);
   }
@@ -1008,10 +781,8 @@ class _$_OnKeypadTapped implements _OnKeypadTapped {
     TResult? Function(_Initial value)? initial,
     TResult? Function(_OnRefresh value)? onRefresh,
     TResult? Function(_ClearPageCommand value)? clearPageCommand,
-    TResult? Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult? Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult? Function(_OnMemoEntered value)? onMemoEntered,
-    TResult? Function(_OnSendTapped value)? onSendTapped,
   }) {
     return onKeypadTapped?.call(this);
   }
@@ -1022,10 +793,8 @@ class _$_OnKeypadTapped implements _OnKeypadTapped {
     TResult Function(_Initial value)? initial,
     TResult Function(_OnRefresh value)? onRefresh,
     TResult Function(_ClearPageCommand value)? clearPageCommand,
-    TResult Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult Function(_OnMemoEntered value)? onMemoEntered,
-    TResult Function(_OnSendTapped value)? onSendTapped,
     required TResult orElse(),
   }) {
     if (onKeypadTapped != null) {
@@ -1035,7 +804,7 @@ class _$_OnKeypadTapped implements _OnKeypadTapped {
   }
 }
 
-abstract class _OnKeypadTapped implements SendEvent {
+abstract class _OnKeypadTapped implements ReceiveEvent {
   const factory _OnKeypadTapped(final KeypadKey tappedKey) = _$_OnKeypadTapped;
 
   KeypadKey get tappedKey;
@@ -1055,7 +824,7 @@ abstract class _$$_OnMemoEnteredCopyWith<$Res> {
 
 /// @nodoc
 class __$$_OnMemoEnteredCopyWithImpl<$Res>
-    extends _$SendEventCopyWithImpl<$Res, _$_OnMemoEntered>
+    extends _$ReceiveEventCopyWithImpl<$Res, _$_OnMemoEntered>
     implements _$$_OnMemoEnteredCopyWith<$Res> {
   __$$_OnMemoEnteredCopyWithImpl(
       _$_OnMemoEntered _value, $Res Function(_$_OnMemoEntered) _then)
@@ -1085,7 +854,7 @@ class _$_OnMemoEntered implements _OnMemoEntered {
 
   @override
   String toString() {
-    return 'SendEvent.onMemoEntered(memo: $memo)';
+    return 'ReceiveEvent.onMemoEntered(memo: $memo)';
   }
 
   @override
@@ -1111,11 +880,8 @@ class _$_OnMemoEntered implements _OnMemoEntered {
     required TResult Function() initial,
     required TResult Function() onRefresh,
     required TResult Function() clearPageCommand,
-    required TResult Function(AmountPercentage amountPercentage)
-        onPercentageTapped,
     required TResult Function(KeypadKey tappedKey) onKeypadTapped,
     required TResult Function(String? memo) onMemoEntered,
-    required TResult Function() onSendTapped,
   }) {
     return onMemoEntered(memo);
   }
@@ -1126,10 +892,8 @@ class _$_OnMemoEntered implements _OnMemoEntered {
     TResult? Function()? initial,
     TResult? Function()? onRefresh,
     TResult? Function()? clearPageCommand,
-    TResult? Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult? Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult? Function(String? memo)? onMemoEntered,
-    TResult? Function()? onSendTapped,
   }) {
     return onMemoEntered?.call(memo);
   }
@@ -1140,10 +904,8 @@ class _$_OnMemoEntered implements _OnMemoEntered {
     TResult Function()? initial,
     TResult Function()? onRefresh,
     TResult Function()? clearPageCommand,
-    TResult Function(AmountPercentage amountPercentage)? onPercentageTapped,
     TResult Function(KeypadKey tappedKey)? onKeypadTapped,
     TResult Function(String? memo)? onMemoEntered,
-    TResult Function()? onSendTapped,
     required TResult orElse(),
   }) {
     if (onMemoEntered != null) {
@@ -1158,10 +920,8 @@ class _$_OnMemoEntered implements _OnMemoEntered {
     required TResult Function(_Initial value) initial,
     required TResult Function(_OnRefresh value) onRefresh,
     required TResult Function(_ClearPageCommand value) clearPageCommand,
-    required TResult Function(_OnPercentageTapped value) onPercentageTapped,
     required TResult Function(_OnKeypadTapped value) onKeypadTapped,
     required TResult Function(_OnMemoEntered value) onMemoEntered,
-    required TResult Function(_OnSendTapped value) onSendTapped,
   }) {
     return onMemoEntered(this);
   }
@@ -1172,10 +932,8 @@ class _$_OnMemoEntered implements _OnMemoEntered {
     TResult? Function(_Initial value)? initial,
     TResult? Function(_OnRefresh value)? onRefresh,
     TResult? Function(_ClearPageCommand value)? clearPageCommand,
-    TResult? Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult? Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult? Function(_OnMemoEntered value)? onMemoEntered,
-    TResult? Function(_OnSendTapped value)? onSendTapped,
   }) {
     return onMemoEntered?.call(this);
   }
@@ -1186,10 +944,8 @@ class _$_OnMemoEntered implements _OnMemoEntered {
     TResult Function(_Initial value)? initial,
     TResult Function(_OnRefresh value)? onRefresh,
     TResult Function(_ClearPageCommand value)? clearPageCommand,
-    TResult Function(_OnPercentageTapped value)? onPercentageTapped,
     TResult Function(_OnKeypadTapped value)? onKeypadTapped,
     TResult Function(_OnMemoEntered value)? onMemoEntered,
-    TResult Function(_OnSendTapped value)? onSendTapped,
     required TResult orElse(),
   }) {
     if (onMemoEntered != null) {
@@ -1199,7 +955,7 @@ class _$_OnMemoEntered implements _OnMemoEntered {
   }
 }
 
-abstract class _OnMemoEntered implements SendEvent {
+abstract class _OnMemoEntered implements ReceiveEvent {
   const factory _OnMemoEntered(final String? memo) = _$_OnMemoEntered;
 
   String? get memo;
@@ -1209,175 +965,38 @@ abstract class _OnMemoEntered implements SendEvent {
 }
 
 /// @nodoc
-abstract class _$$_OnSendTappedCopyWith<$Res> {
-  factory _$$_OnSendTappedCopyWith(
-          _$_OnSendTapped value, $Res Function(_$_OnSendTapped) then) =
-      __$$_OnSendTappedCopyWithImpl<$Res>;
-}
-
-/// @nodoc
-class __$$_OnSendTappedCopyWithImpl<$Res>
-    extends _$SendEventCopyWithImpl<$Res, _$_OnSendTapped>
-    implements _$$_OnSendTappedCopyWith<$Res> {
-  __$$_OnSendTappedCopyWithImpl(
-      _$_OnSendTapped _value, $Res Function(_$_OnSendTapped) _then)
-      : super(_value, _then);
-}
-
-/// @nodoc
-
-class _$_OnSendTapped implements _OnSendTapped {
-  const _$_OnSendTapped();
-
-  @override
-  String toString() {
-    return 'SendEvent.onSendTapped()';
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is _$_OnSendTapped);
-  }
-
-  @override
-  int get hashCode => runtimeType.hashCode;
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() onRefresh,
-    required TResult Function() clearPageCommand,
-    required TResult Function(AmountPercentage amountPercentage)
-        onPercentageTapped,
-    required TResult Function(KeypadKey tappedKey) onKeypadTapped,
-    required TResult Function(String? memo) onMemoEntered,
-    required TResult Function() onSendTapped,
-  }) {
-    return onSendTapped();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? onRefresh,
-    TResult? Function()? clearPageCommand,
-    TResult? Function(AmountPercentage amountPercentage)? onPercentageTapped,
-    TResult? Function(KeypadKey tappedKey)? onKeypadTapped,
-    TResult? Function(String? memo)? onMemoEntered,
-    TResult? Function()? onSendTapped,
-  }) {
-    return onSendTapped?.call();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? onRefresh,
-    TResult Function()? clearPageCommand,
-    TResult Function(AmountPercentage amountPercentage)? onPercentageTapped,
-    TResult Function(KeypadKey tappedKey)? onKeypadTapped,
-    TResult Function(String? memo)? onMemoEntered,
-    TResult Function()? onSendTapped,
-    required TResult orElse(),
-  }) {
-    if (onSendTapped != null) {
-      return onSendTapped();
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(_Initial value) initial,
-    required TResult Function(_OnRefresh value) onRefresh,
-    required TResult Function(_ClearPageCommand value) clearPageCommand,
-    required TResult Function(_OnPercentageTapped value) onPercentageTapped,
-    required TResult Function(_OnKeypadTapped value) onKeypadTapped,
-    required TResult Function(_OnMemoEntered value) onMemoEntered,
-    required TResult Function(_OnSendTapped value) onSendTapped,
-  }) {
-    return onSendTapped(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(_Initial value)? initial,
-    TResult? Function(_OnRefresh value)? onRefresh,
-    TResult? Function(_ClearPageCommand value)? clearPageCommand,
-    TResult? Function(_OnPercentageTapped value)? onPercentageTapped,
-    TResult? Function(_OnKeypadTapped value)? onKeypadTapped,
-    TResult? Function(_OnMemoEntered value)? onMemoEntered,
-    TResult? Function(_OnSendTapped value)? onSendTapped,
-  }) {
-    return onSendTapped?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(_Initial value)? initial,
-    TResult Function(_OnRefresh value)? onRefresh,
-    TResult Function(_ClearPageCommand value)? clearPageCommand,
-    TResult Function(_OnPercentageTapped value)? onPercentageTapped,
-    TResult Function(_OnKeypadTapped value)? onKeypadTapped,
-    TResult Function(_OnMemoEntered value)? onMemoEntered,
-    TResult Function(_OnSendTapped value)? onSendTapped,
-    required TResult orElse(),
-  }) {
-    if (onSendTapped != null) {
-      return onSendTapped(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class _OnSendTapped implements SendEvent {
-  const factory _OnSendTapped() = _$_OnSendTapped;
-}
-
-/// @nodoc
-mixin _$SendState {
+mixin _$ReceiveState {
   PageState get pageState => throw _privateConstructorUsedError;
   PageCommand? get command => throw _privateConstructorUsedError;
   String? get userEnteredAmount => throw _privateConstructorUsedError;
   String? get memo => throw _privateConstructorUsedError;
-  UserProfileData get receiverUser => throw _privateConstructorUsedError;
   WalletTokenData get tokenData => throw _privateConstructorUsedError;
-  bool get isSendButtonLoading => throw _privateConstructorUsedError;
 
   @JsonKey(ignore: true)
-  $SendStateCopyWith<SendState> get copyWith =>
+  $ReceiveStateCopyWith<ReceiveState> get copyWith =>
       throw _privateConstructorUsedError;
 }
 
 /// @nodoc
-abstract class $SendStateCopyWith<$Res> {
-  factory $SendStateCopyWith(SendState value, $Res Function(SendState) then) =
-      _$SendStateCopyWithImpl<$Res, SendState>;
+abstract class $ReceiveStateCopyWith<$Res> {
+  factory $ReceiveStateCopyWith(
+          ReceiveState value, $Res Function(ReceiveState) then) =
+      _$ReceiveStateCopyWithImpl<$Res, ReceiveState>;
   @useResult
   $Res call(
       {PageState pageState,
       PageCommand? command,
       String? userEnteredAmount,
       String? memo,
-      UserProfileData receiverUser,
-      WalletTokenData tokenData,
-      bool isSendButtonLoading});
+      WalletTokenData tokenData});
 
   $PageCommandCopyWith<$Res>? get command;
-  $UserProfileDataCopyWith<$Res> get receiverUser;
 }
 
 /// @nodoc
-class _$SendStateCopyWithImpl<$Res, $Val extends SendState>
-    implements $SendStateCopyWith<$Res> {
-  _$SendStateCopyWithImpl(this._value, this._then);
+class _$ReceiveStateCopyWithImpl<$Res, $Val extends ReceiveState>
+    implements $ReceiveStateCopyWith<$Res> {
+  _$ReceiveStateCopyWithImpl(this._value, this._then);
 
   // ignore: unused_field
   final $Val _value;
@@ -1391,9 +1010,7 @@ class _$SendStateCopyWithImpl<$Res, $Val extends SendState>
     Object? command = freezed,
     Object? userEnteredAmount = freezed,
     Object? memo = freezed,
-    Object? receiverUser = null,
     Object? tokenData = null,
-    Object? isSendButtonLoading = null,
   }) {
     return _then(_value.copyWith(
       pageState: null == pageState
@@ -1412,18 +1029,10 @@ class _$SendStateCopyWithImpl<$Res, $Val extends SendState>
           ? _value.memo
           : memo // ignore: cast_nullable_to_non_nullable
               as String?,
-      receiverUser: null == receiverUser
-          ? _value.receiverUser
-          : receiverUser // ignore: cast_nullable_to_non_nullable
-              as UserProfileData,
       tokenData: null == tokenData
           ? _value.tokenData
           : tokenData // ignore: cast_nullable_to_non_nullable
               as WalletTokenData,
-      isSendButtonLoading: null == isSendButtonLoading
-          ? _value.isSendButtonLoading
-          : isSendButtonLoading // ignore: cast_nullable_to_non_nullable
-              as bool,
     ) as $Val);
   }
 
@@ -1438,21 +1047,14 @@ class _$SendStateCopyWithImpl<$Res, $Val extends SendState>
       return _then(_value.copyWith(command: value) as $Val);
     });
   }
-
-  @override
-  @pragma('vm:prefer-inline')
-  $UserProfileDataCopyWith<$Res> get receiverUser {
-    return $UserProfileDataCopyWith<$Res>(_value.receiverUser, (value) {
-      return _then(_value.copyWith(receiverUser: value) as $Val);
-    });
-  }
 }
 
 /// @nodoc
-abstract class _$$_SendStateCopyWith<$Res> implements $SendStateCopyWith<$Res> {
-  factory _$$_SendStateCopyWith(
-          _$_SendState value, $Res Function(_$_SendState) then) =
-      __$$_SendStateCopyWithImpl<$Res>;
+abstract class _$$_ReceiveStateCopyWith<$Res>
+    implements $ReceiveStateCopyWith<$Res> {
+  factory _$$_ReceiveStateCopyWith(
+          _$_ReceiveState value, $Res Function(_$_ReceiveState) then) =
+      __$$_ReceiveStateCopyWithImpl<$Res>;
   @override
   @useResult
   $Res call(
@@ -1460,22 +1062,18 @@ abstract class _$$_SendStateCopyWith<$Res> implements $SendStateCopyWith<$Res> {
       PageCommand? command,
       String? userEnteredAmount,
       String? memo,
-      UserProfileData receiverUser,
-      WalletTokenData tokenData,
-      bool isSendButtonLoading});
+      WalletTokenData tokenData});
 
   @override
   $PageCommandCopyWith<$Res>? get command;
-  @override
-  $UserProfileDataCopyWith<$Res> get receiverUser;
 }
 
 /// @nodoc
-class __$$_SendStateCopyWithImpl<$Res>
-    extends _$SendStateCopyWithImpl<$Res, _$_SendState>
-    implements _$$_SendStateCopyWith<$Res> {
-  __$$_SendStateCopyWithImpl(
-      _$_SendState _value, $Res Function(_$_SendState) _then)
+class __$$_ReceiveStateCopyWithImpl<$Res>
+    extends _$ReceiveStateCopyWithImpl<$Res, _$_ReceiveState>
+    implements _$$_ReceiveStateCopyWith<$Res> {
+  __$$_ReceiveStateCopyWithImpl(
+      _$_ReceiveState _value, $Res Function(_$_ReceiveState) _then)
       : super(_value, _then);
 
   @pragma('vm:prefer-inline')
@@ -1485,11 +1083,9 @@ class __$$_SendStateCopyWithImpl<$Res>
     Object? command = freezed,
     Object? userEnteredAmount = freezed,
     Object? memo = freezed,
-    Object? receiverUser = null,
     Object? tokenData = null,
-    Object? isSendButtonLoading = null,
   }) {
-    return _then(_$_SendState(
+    return _then(_$_ReceiveState(
       pageState: null == pageState
           ? _value.pageState
           : pageState // ignore: cast_nullable_to_non_nullable
@@ -1506,33 +1102,23 @@ class __$$_SendStateCopyWithImpl<$Res>
           ? _value.memo
           : memo // ignore: cast_nullable_to_non_nullable
               as String?,
-      receiverUser: null == receiverUser
-          ? _value.receiverUser
-          : receiverUser // ignore: cast_nullable_to_non_nullable
-              as UserProfileData,
       tokenData: null == tokenData
           ? _value.tokenData
           : tokenData // ignore: cast_nullable_to_non_nullable
               as WalletTokenData,
-      isSendButtonLoading: null == isSendButtonLoading
-          ? _value.isSendButtonLoading
-          : isSendButtonLoading // ignore: cast_nullable_to_non_nullable
-              as bool,
     ));
   }
 }
 
 /// @nodoc
 
-class _$_SendState extends _SendState {
-  const _$_SendState(
+class _$_ReceiveState extends _ReceiveState {
+  const _$_ReceiveState(
       {this.pageState = PageState.initial,
       this.command,
       this.userEnteredAmount = null,
       this.memo = null,
-      required this.receiverUser,
-      required this.tokenData,
-      this.isSendButtonLoading = false})
+      required this.tokenData})
       : super._();
 
   @override
@@ -1547,58 +1133,47 @@ class _$_SendState extends _SendState {
   @JsonKey()
   final String? memo;
   @override
-  final UserProfileData receiverUser;
-  @override
   final WalletTokenData tokenData;
-  @override
-  @JsonKey()
-  final bool isSendButtonLoading;
 
   @override
   String toString() {
-    return 'SendState(pageState: $pageState, command: $command, userEnteredAmount: $userEnteredAmount, memo: $memo, receiverUser: $receiverUser, tokenData: $tokenData, isSendButtonLoading: $isSendButtonLoading)';
+    return 'ReceiveState(pageState: $pageState, command: $command, userEnteredAmount: $userEnteredAmount, memo: $memo, tokenData: $tokenData)';
   }
 
   @override
   bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _$_SendState &&
+            other is _$_ReceiveState &&
             (identical(other.pageState, pageState) ||
                 other.pageState == pageState) &&
             (identical(other.command, command) || other.command == command) &&
             (identical(other.userEnteredAmount, userEnteredAmount) ||
                 other.userEnteredAmount == userEnteredAmount) &&
             (identical(other.memo, memo) || other.memo == memo) &&
-            (identical(other.receiverUser, receiverUser) ||
-                other.receiverUser == receiverUser) &&
             (identical(other.tokenData, tokenData) ||
-                other.tokenData == tokenData) &&
-            (identical(other.isSendButtonLoading, isSendButtonLoading) ||
-                other.isSendButtonLoading == isSendButtonLoading));
+                other.tokenData == tokenData));
   }
 
   @override
-  int get hashCode => Object.hash(runtimeType, pageState, command,
-      userEnteredAmount, memo, receiverUser, tokenData, isSendButtonLoading);
+  int get hashCode => Object.hash(
+      runtimeType, pageState, command, userEnteredAmount, memo, tokenData);
 
   @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
-  _$$_SendStateCopyWith<_$_SendState> get copyWith =>
-      __$$_SendStateCopyWithImpl<_$_SendState>(this, _$identity);
+  _$$_ReceiveStateCopyWith<_$_ReceiveState> get copyWith =>
+      __$$_ReceiveStateCopyWithImpl<_$_ReceiveState>(this, _$identity);
 }
 
-abstract class _SendState extends SendState {
-  const factory _SendState(
+abstract class _ReceiveState extends ReceiveState {
+  const factory _ReceiveState(
       {final PageState pageState,
       final PageCommand? command,
       final String? userEnteredAmount,
       final String? memo,
-      required final UserProfileData receiverUser,
-      required final WalletTokenData tokenData,
-      final bool isSendButtonLoading}) = _$_SendState;
-  const _SendState._() : super._();
+      required final WalletTokenData tokenData}) = _$_ReceiveState;
+  const _ReceiveState._() : super._();
 
   @override
   PageState get pageState;
@@ -1609,13 +1184,9 @@ abstract class _SendState extends SendState {
   @override
   String? get memo;
   @override
-  UserProfileData get receiverUser;
-  @override
   WalletTokenData get tokenData;
   @override
-  bool get isSendButtonLoading;
-  @override
   @JsonKey(ignore: true)
-  _$$_SendStateCopyWith<_$_SendState> get copyWith =>
+  _$$_ReceiveStateCopyWith<_$_ReceiveState> get copyWith =>
       throw _privateConstructorUsedError;
 }
