+++ lib/ui/transfer_tokens/receive/interactor/receive_event.dart
