+++ lib/ui/transfer_tokens/send/components/send_review_bottom_sheet.dart
@@ -10,8 +10,8 @@ import 'package:hypha_wallet/design/circle_with_icon/color_arrow_up.dart';
 import 'package:hypha_wallet/design/hypha_card.dart';
 import 'package:hypha_wallet/design/hypha_colors.dart';
 import 'package:hypha_wallet/design/themes/extensions/theme_extension_provider.dart';
-import 'package:hypha_wallet/ui/send/components/send_to_user_row.dart';
-import 'package:hypha_wallet/ui/send/interactor/send_bloc.dart';
+import 'package:hypha_wallet/ui/transfer_tokens/send/components/send_to_user_row.dart';
+import 'package:hypha_wallet/ui/transfer_tokens/send/interactor/send_bloc.dart';
 
 class SendReviewBottomSheet extends StatelessWidget {
   const SendReviewBottomSheet({super.key});
