+++ lib/ui/transfer_tokens/send/components/send_view.dart
@@ -1,19 +1,18 @@
-import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:get/get.dart';
 import 'package:hypha_wallet/design/background/hypha_page_background.dart';
 import 'package:hypha_wallet/design/bottom_component/hypha_safe_bottom_navigation_bar.dart';
 import 'package:hypha_wallet/design/buttons/button_type.dart';
 import 'package:hypha_wallet/design/buttons/hypha_app_button.dart';
-import 'package:hypha_wallet/design/hypha_card.dart';
 import 'package:hypha_wallet/design/hypha_colors.dart';
 import 'package:hypha_wallet/design/themes/extensions/theme_extension_provider.dart';
-import 'package:hypha_wallet/ui/send/components/send_memo_field.dart';
-import 'package:hypha_wallet/ui/send/components/send_review_bottom_sheet.dart';
-import 'package:hypha_wallet/ui/send/components/send_to_user_row.dart';
-import 'package:hypha_wallet/ui/send/data/amount_percentage.dart';
-import 'package:hypha_wallet/ui/send/data/keypad_key.dart';
-import 'package:hypha_wallet/ui/send/interactor/send_bloc.dart';
+import 'package:hypha_wallet/ui/transfer_tokens/components/memo_field.dart';
+import 'package:hypha_wallet/ui/transfer_tokens/components/number_keyboard_grid.dart';
+import 'package:hypha_wallet/ui/transfer_tokens/send/components/send_review_bottom_sheet.dart';
+import 'package:hypha_wallet/ui/transfer_tokens/send/components/send_to_user_row.dart';
+import 'package:hypha_wallet/ui/transfer_tokens/send/data/amount_percentage.dart';
+import 'package:hypha_wallet/ui/transfer_tokens/send/interactor/send_bloc.dart';
 
 class SendView extends StatelessWidget {
   const SendView({super.key});
@@ -70,53 +69,29 @@ class SendView extends StatelessWidget {
               const _AvailableBalanceWidget(),
               const SendToUserRow(imageRadius: 20),
               const SizedBox(height: 24),
-              const SendMemoField(),
+              BlocBuilder<SendBloc, SendState>(
+                buildWhen: (c, p) => c.memo != p.memo,
+                builder: (context, state) {
+                  return MemoField(
+                    onPressed: (value) {
+                      context.read<SendBloc>().add(SendEvent.onMemoEntered(value));
+                      Get.back();
+                    },
+                    memo: state.memo,
+                  );
+                },
+              ),
               const SizedBox(height: 24),
               const _PercentagesWidget(),
               const SizedBox(height: 24),
-              const NumberKeyboardGrid(),
+              NumberKeyboardGrid(onKeyTapped: (element) {
+                context.read<SendBloc>().add(SendEvent.onKeypadTapped(element));
+              }),
             ],
           ),
         ),
       ),
     );
-  }
-}
-
-class NumberKeyboardGrid extends StatelessWidget {
-  const NumberKeyboardGrid({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return GridView.count(
-        crossAxisCount: 3,
-        shrinkWrap: true,
-        childAspectRatio: 1.8,
-        physics: const NeverScrollableScrollPhysics(),
-        children: KeypadKey.values.mapIndexed((index, element) {
-          final text = switch (element) {
-            KeypadKey.dot => '.',
-            KeypadKey.delete => '<-',
-            _ => element.value.toString(),
-          };
-          return GestureDetector(
-            onTap: () {
-              context.read<SendBloc>().add(SendEvent.onKeypadTapped(element));
-            },
-            child: Padding(
-              padding: EdgeInsets.only(top: 8, bottom: 8, left: index % 3 == 0 ? 0 : 8, right: index % 3 == 2 ? 0 : 8),
-              child: HyphaCard(
-                borderRadius: BorderRadius.circular(14),
-                child: Center(
-                  child: Text(
-                    text,
-                    style: context.hyphaTextTheme.mediumTitles,
-                  ),
-                ),
-              ),
-            ),
-          );
-        }).toList());
   }
 }
