+++ lib/ui/transfer_tokens/send/data/amount_percentage.dart
